app/requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
blinker==1.4
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
enum34==1.1.6
env==0.1.0
Flask==0.10.1
Flask-Images==2.1.2
Flask-KVSession==0.6.2
Flask-Login==0.4.1
Flask-OAuth==0.12
flask-oauth2-login==0.0.9
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
identity-toolkit-python-client==0.1.11
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
linecache2==1.0.0
MarkupSafe==1.0
marshmallow==2.15.3
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.1.0
PIL==1.1.7
Pillow==5.1.0
pillowcase==2.0.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodomex==3.6.2
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-gflags==2.0
pytz==2018.4
rauth==0.7.3
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
simplejson==3.15.0
simplekv==0.9.2
six==1.11.0
SQLAlchemy==1.2.9
traceback2==1.4.0
Unidecode==1.0.22
unittest2==1.1.0
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.23
virtualenv==16.0.0
webargs==3.0.1
Werkzeug==0.14.1
WTForms==2.2.1
